Shuffle some stuff around
